refactor : WinningLottoTicket 에서 LottoTicket 대신 List<LottoNo> 를 갖도록 변경
당첨 복권은 당첨 티켓이 아닌 번호를 갖으므로 List<LottoNo> 로 변경하고, 당첨 복권 여부 확인 메서드를 WinningLottoTicket에서 처리하도록 변경
